+++ src/archivist/errors.js
@@ -1,9 +1,9 @@
 export class InaccessibleContentError extends Error {
-  constructor(message) {
-    if (Array.isArray(message)) {
-      message = `\n - ${message.join('\n - ')}`;
-    }
-    super(`The documents cannot be accessed or their contents can not be selected:${message}`);
+  constructor(reasonOrReasons) {
+    const reasons = [].concat(reasonOrReasons);
+
+    super(`The documents cannot be accessed or their contents can not be selected:${`\n - ${reasons.join('\n - ')}`}`);
     this.name = 'InaccessibleContentError';
+    this.reasons = reasons;
   }
 }